Change in copier default behavior
- I think this tries to do the right thing, but if shutil tries
  to remove anything with resource forks it will fail, because
  shutil cannot deal with xattrs in python 2.x...

+++ Code/autopkglib/Copier.py
@@ -24,6 +24,8 @@ from autopkglib.DmgMounter import DmgMounter
 
 __all__ = ["Copier"]
 
+OVERWRITE_DEFAULT = True
+
 
 class Copier(DmgMounter):
     """Copies source_path to destination_path."""
@@ -44,8 +46,8 @@ class Copier(DmgMounter):
         "overwrite": {
             "required": False,
             "description": (
-                "Whether the destination will be overwritten if necessary. "
-                "Uses a boolean value."
+                "Boolean for whether the destination will be overwritten "
+                "if it already exists. Defaults to %s." % OVERWRITE_DEFAULT
             ),
         },
     }
@@ -66,21 +68,23 @@ class Copier(DmgMounter):
                 raise ProcessorError("Can't remove %s: %s" % (dest_item, err.strerror))
 
         # Copy file or directory.
-        try:
-            if os.path.isdir(source_item):
-                shutil.copytree(source_item, dest_item, symlinks=True)
-            elif not os.path.isdir(dest_item):
-                shutil.copyfile(source_item, dest_item)
-            else:
-                shutil.copy(source_item, dest_item)
-            self.output("Copied %s to %s" % (source_item, dest_item))
-        except BaseException as err:
-            raise ProcessorError(
-                "Can't copy %s to %s: %s" % (source_item, dest_item, err)
-            )
+        if not os.path.exists(dest_item) or overwrite:
+            try:
+                if os.path.isdir(source_item):
+                    shutil.copytree(source_item, dest_item, symlinks=True)
+                elif not os.path.isdir(dest_item):
+                    shutil.copyfile(source_item, dest_item)
+                else:
+                    shutil.copy(source_item, dest_item)
+                self.output("Copied %s to %s" % (source_item, dest_item))
+            except BaseException as err:
+                raise ProcessorError(
+                    "Can't copy %s to %s: %s" % (source_item, dest_item, err)
+                )
 
     def main(self):
         source_path = self.env["source_path"]
+        overwrite = self.env.get("overwrite", OVERWRITE_DEFAULT)
         # Check if we're trying to copy something inside a dmg.
         (dmg_path, dmg, dmg_source_path) = self.parsePathForDMG(source_path)
         try:
@@ -110,11 +114,7 @@ class Copier(DmgMounter):
                 )
 
             # do the copy
-            self.copy(
-                matched_source_path,
-                self.env["destination_path"],
-                overwrite=self.env.get("overwrite"),
-            )
+            self.copy(matched_source_path, self.env["destination_path"], overwrite)
         finally:
             if dmg:
                 self.unmount(dmg_path)
